Commit: undefinedbnwIPoUqMr

+++ backend/src/chat/chat.service.ts
@@ -407,7 +407,7 @@ export class ChatService {
 						groupId: group.id
 					}
 				}
-			})
+			});
 			if (userGroup) {
 				await this.prisma.userGROUP.delete({
 					where: {
@@ -417,15 +417,26 @@ export class ChatService {
 						}
 					}
 				});
-				if (chatHistory)
-					await this.prisma.userCHATHISTORY.delete({
-						where: {
-							userid_chathistoryid: {
-								userid: user.id,
-								chathistoryid: chatHistory.id
+				if (chatHistory) {
+					const	userchatHistory = await this.
+						prisma.userCHATHISTORY.findUnique({
+							where: {
+								userid_chathistoryid: {
+									userid: user.id,
+									chathistoryid: chatHistory.id
+								}
 							}
-						}
 					});
+					if (userchatHistory)
+						await this.prisma.userCHATHISTORY.delete({
+							where: {
+								userid_chathistoryid: {
+									userid: user.id,
+									chathistoryid: chatHistory.id
+								}
+							}
+						});
+				}
 				
 			}
 			else {
@@ -650,12 +661,70 @@ export class ChatService {
 		const	group = await this.prisma.gROUP.findFirst({
 			where: { name: data.name }
 		})
-		return [{
-			id: group.id,
-			name: group.name,
-			password: group.hash ? true : false,
-			banded: group.banded,
-			muted: group.muted
-		}];
+		if (group) {
+			return [{
+				id: group.id,
+				name: group.name,
+				password: group.hash ? true : false,
+				banded: group.banded,
+				muted: group.muted
+			}];
+		}
+		return [];
+	}
+	async groupsChage(data: {
+		name: string,
+		old: string,
+		password: string,
+		oldName: string
+	}) {
+		const	oldGroup = await this.prisma.gROUP.findFirst({
+			where: { name: data.oldName }
+		});
+		if (oldGroup) {
+			if (data.name.length) {
+				const	group = await this.prisma.gROUP.findFirst({
+					where: { name: data.name }
+				});
+				if (group)
+					return "wrongUser";
+				await this.prisma.gROUP.update({
+					where: { id: oldGroup.id },
+					data: { name: data.name }
+				});
+				await this.prisma.cHATHISTORY.update({
+					where: { name: oldGroup.name },
+					data: { name: data.name }
+				});
+			}
+			if (data.old.length) {
+				const	match = await compare(data.old, oldGroup.hash);
+				if (!match)
+					return "wrongPassword";
+				let	hashed : string | null;
+				if (data.password.length)
+					hashed = await hash(data.password, 10);
+				else
+					hashed = null;
+				await this.prisma.gROUP.update({
+					where: { id: oldGroup.id },
+					data: { hash: hashed }
+				});
+			}
+			else if (data.password.length)
+				await this.prisma.gROUP.update({
+					where: { id: oldGroup.id },
+					data: { hash: await hash(data.password, 10) }
+				});
+			return "DONE";
+		}
+	}
+	async checkGroup(data: { name: string }) {
+		const	group = await this.prisma.gROUP.findFirst({
+			where: { name: data.name }
+		});
+		if (group)
+			return true;
+		return false
 	}
 }
